feat(#5 backend): successfully getting the access token
