+++ package.json
@@ -14,13 +14,12 @@
   "dependencies": {
     "@types/express": "^4.17.17",
     "@types/node": "^20.3.1",
-    "@types/spotify-web-api-node": "^5.0.7",
+    "@types/request": "^2.48.8",
     "@types/superagent": "^4.1.18",
     "dotenv": "^16.1.4",
     "eslint": "^8.42.0",
     "express": "^4.18.2",
     "querystring": "^0.2.1",
-    "spotify-web-api-node": "^5.0.2",
     "superagent": "^8.0.9",
     "typescript": "^5.1.3"
   },
