--- src/utils/initSpotifyApi.ts
@@ -1,12 +1,0 @@
-import SpotifyWebApi from "spotify-web-api-node";
-import secrets from "../secrets";
-
-function initSpotifyApi() {
-  return new SpotifyWebApi({
-    clientId: secrets.clientId,
-    clientSecret: secrets.clientSecret,
-    redirectUri: secrets.redirectUri,
-  });
-}
-
-export default initSpotifyApi;
